refactor: vite 기반 프로젝트 구조 리팩터링
