--- config.dev.js
@@ -1,9 +1,0 @@
-const CONFIG = {
-  BASE_URL: 'http://localhost:8080',
-  ENV: 'development',
-  STORAGE_KEYS: {
-    EMAIL: 'email',
-    IDENTIFIER: 'identifier',
-    IS_AUTHENTICATED: 'isAuthenticated'
-  }
-};
