--- config.prod.js
@@ -1,9 +1,0 @@
-const CONFIG = {
-  BASE_URL: 'https://api.your-domain.com',
-  ENV: 'production',
-  STORAGE_KEYS: {
-    EMAIL: 'email',
-    IDENTIFIER: 'identifier',
-    IS_AUTHENTICATED: 'isAuthenticated'
-  }
-};
