+++ public/popup.html
@@ -71,7 +71,6 @@
     </div>
   </div>
 
-  <script src="config.js"></script>
-  <script src="popup.js"></script>
+  <script type="module" src="popup.js"></script>
 </body>
 </html>
