+++ vite.config.js
@@ -1,0 +1,26 @@
+import { defineConfig } from 'vite';
+import { resolve } from 'path';
+
+export default defineConfig({
+  publicDir: 'public',
+  build: {
+    outDir: 'dist',
+    emptyDirFirst: true,
+    rollupOptions: {
+      input: {
+        popup: resolve(__dirname, 'src/popup.js'),
+        background: resolve(__dirname, 'src/background.js')
+      },
+      output: {
+        entryFileNames: '[name].js',
+        chunkFileNames: 'chunks/[name].js',
+        assetFileNames: '[name].[ext]'
+      }
+    }
+  },
+  resolve: {
+    alias: {
+      '@': resolve(__dirname, 'src')
+    }
+  }
+});
